wip: feature function #25 - RMLFNOTC0005-CSV working
There were some issues with finding the logical source. Also a `student.csv` is used, however, this file differs from the others regarding the use of a url-column. 
To make the distinction clear, I added a `student_url.csv` to the `fno-testcases` in the working testcases dir, as it appears to look for its logical sources in that dir. 

engine.statement.SubjectGeneratorAssembler
-  invokes FunctionMapGeneratorAssembler when the subjectmap has a FunctionMap

test:
- moved test from failing to working dir
- created `student_url.csv` in the working `fno-testcases` dir. 
- updated RMLFNOTC0005-CSV/mapping.ttl's logical source  to point at `fno-testcases/student_url.csv`