CR: MappingExtractor, unused cls argument
RQ: Why has cls the Any type if it's supposed to be a mapper? And is that argument used at all?
RA: The `cls` argument seems to be pointless, actually. 
CR: unused FunctionMapExtractor argument

RQ: Is there ever a FunctionMapExtractor given as argument, or would that be so in the future? If not, maybe add it as constructor in the next line ... 
RA: I think I added this with the idea to support both old and new FnO versions. The argument can be removed as it isn't really used.
CR: adding classname to debug statements isn't necessary

RQ: A general comment to statements like this one: is it necessary to add the classname? I guess log4j can be configured to print it..
RA: Indeed, I wasn't aware of that when I was writing that :p 
CR: analoguous to ff9443a34a843f2ee6de86cb58a6f9a620100468


CR: checking whether a resource is a Triples Map

RQ: Be careful here: a predicate-object map is not required for a triples map, according to the spec. So if this code is used to check if a mapping is correct, then the code is wrong. If the code is only used to check wheter triples will be generated, then this is correct behaviour (since no triples will be generated if no predicate-object map is present, so this triples map can be omitted). However, please adjust the comments to clarify if the latter is the case.
RA: The idea was to filter out triples maps resources from a given RDFGraph because, initially, the method `filterTriplesMaps` filtered triples maps in a graph based on two conditions; 1. the resource's class is a TriplesMap, 2. the resource has a logical source property. 

```scala
val typeUri = Uri(RMLVoc.Class.TRIPLESMAP)
val triplesMapResources = (graph.filterResources(typeUri) ++
      graph.filterProperties(Uri(RMLVoc.Property.LOGICALSOURCE))).distinct
```
However, this resulted in errors when processing a Function Map, because a Function Map can also have a logical source property defined. In a later processing step this would cause errors because the RML Streamer would try to extract a subject map from the so-called Triples Map.

> In the case of [r2]rml, Term Maps determine how to generate an rdf term
relying on references to raw data. Therefore, a new type of Term Map was introduced, the Function Map (`fnml:FunctionMap`, Listing 3: line 10). A Function Map is a Term Map generated by executing a function, instead of using a constant or a reference to the raw data values. In contrast to an RDF Term Map that uses values referenced from a Logical Source to generate an rdf term, a Function Map uses values referenced from a Logical Source to execute a function. Article: 
[Declarative Data Transformations for
Linked Data Generation: the case of DBpedia](https://biblio.ugent.be/publication/8525863/file/8525869)

Therefore, I changed `filterTriplesMaps` to do the following
1. filter out resources having a logical source (these are the `potentialTriplesMapResources`)
2. filter out resources that meet the requirements of a Triples Maps defined by the spec
```scala
val logicalSourcePropertyUri = Uri(RMLVoc.Property.LOGICALSOURCE)
val potentialTriplesMapResources = graph.filterProperties(logicalSourcePropertyUri)
val triplesMapResources = potentialTriplesMapResources.filter(isTriplesMap)
```


However, as you suggested, the check in `isTriplesMap` regarding the predicate-object maps was not correct.
You are right about the predicate-object maps. A triples map can indeed contain zero or more predicate-object maps.

Fix: `isTriplesMap` doesn't check the presence of predicate-object map resources because a triples map can contain zero or more.
Furthermore, the checks for logical source and subject subjectmap are now more stringent by checking whether there's exactly one of both.  Also, the shortcut for subjectmap is also being detected now.
CR: removed unused commented code


CR: removal of xsd:number

RQ: If XML Schema doesn't define a xsd:number then remove it. Maybe grel.ttl needs an update. Or maybe a number is defined in another namespace? Anyway, I think number and decimal are the same in this case.


CR: unused class